requirements.txt
pika
boto3
pandas
pyarrow
python-dotenv
requests